reorganise iiif-builder modules
