+++ src/iiif-builder/requirements.txt
@@ -10,3 +10,4 @@ microsoft-kiota-serialization-text~=1.9.2
 psycopg~=3.2.6
 lxml~=5.3.1
 msal~=1.32.0
+python-dotenv~=1.0.1
